вынес запросы в БД в отдельный файл
